apps/api/requirements.txt
fastapi
uvicorn
jinja2
google-cloud-firestore>=2.15.0

